Refine the authorization system by allowing limited public read access. Closes #30.
Copying and pasting from the comments for authz.py, we find...

    There are 2 basic authorization scenarios:

    1) Public request: no user provided

    2) Registered request: user known and (later) authenticated

    In scenario 1, we allow the action if BOTH of the following criteria are
    satisfied, namely a) the annotation has a null or empty permissions field
    for that action, AND b) the action is in the PUBLIC_ACTIONS list.

    In scenario 2, we allow the action if EITHER of the following criteria are
    satisfied, namely a) the annotation has a null or empty permissions field
    for that action, OR b) the user is listed in the permissions field for
    that action.

In addition, this commit adds proper authorization checking to the search
APIs. As a guest user, you will only ever get annotations with null or empty
'read' permissions fields. As an authenticated user, you will get annotations
which have either null/empty 'read' fields, or 'read' fields in which you are
included.