Simplify state management within ElasticSearch ext
Rather than explicitly looking up the app within the ElasticSearch Flask
extension, we can lean more heavily on Flask, and just use current_app
to do this.

We also ensure that a connection instance is bound to a given context,
and is memoised for the duration of the context's life. This means that
connection pooling within the connection class is respected, and we
don't open a new TCP connection every time we call es.conn.